+++ example/lib/main.dart
@@ -14,6 +14,7 @@ class MyApp extends StatelessWidget {
       title: 'StringContains Example',
       theme: ThemeData(
         primarySwatch: Colors.blue,
+        useMaterial3: true,
       ),
       home: const MyHomePage(),
     );
@@ -28,7 +29,7 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  final stringContainsBadWords = 'I am CEO, Bitch';
+  final String? stringContainsBadWords = 'I am CEO, Bitch';
   final stringContainsUrl = 'Please visit our website : https://betterx.io/.';
   final stringContainsMultipleUrls =
       'Please visit our website : https://betterx.io/ and you can search for any thing on google : google.com, I love Dart : https://dart.dev/, you can practice on : https://dartpad.dev/. Flutter uses Dart, you can create your own app with Flutter : https://flutter.dev/, this is a good example of Flutter : https://flutter.dev/docs/get-started/.';
@@ -38,6 +39,8 @@ class _MyHomePageState extends State<MyHomePage> {
   final normalString = 'I love Dart and Flutter';
   final stringContainsHashtags =
       'I love #dart and #flutter. @flutter is #awesome';
+
+  String? stringOfBadWords = 'I am ceo bitch';
 
   @override
   Widget build(BuildContext context) {
@@ -91,6 +94,13 @@ class _MyHomePageState extends State<MyHomePage> {
                 // hide email
                 Text(
                   stringContainsEmail.hideEmails(),
+                  textAlign: TextAlign.center,
+                  style: Theme.of(context).textTheme.titleLarge?.copyWith(
+                        color: Theme.of(context).primaryColor,
+                      ),
+                ),
+                Text(
+                  stringOfBadWords.cleanBadWords(),
                   textAlign: TextAlign.center,
                   style: Theme.of(context).textTheme.titleLarge?.copyWith(
                         color: Theme.of(context).primaryColor,
@@ -158,13 +168,68 @@ class _MyHomePageState extends State<MyHomePage> {
                   isSelecatable: true,
                   source:
                       '$stringContainsBadWords\n$stringContainsUrl\n$stringContainsMultipleUrls\n$stringContainsEmail\n$stringContainsPhoneNumber\n$normalString\n$stringContainsHashtags',
-                  onTap: (url) {
-                    print('element is a ${url.value} and type of ${url.type}');
-                    if (url.type == StringContainsElementType.words) {
+                  onTap: (element) {
+                    print(
+                        'element is a ${element.value} and type of ${element.type}');
+                    if (element.type == StringContainsElementType.words) {
                       ScaffoldMessenger.of(context).showSnackBar(
                         SnackBar(
                           content: Text(
-                            '${url.value} is awesome!!',
+                            '${element.value} is awesome!!',
+                          ),
+                          duration: const Duration(seconds: 2),
+                          dismissDirection: DismissDirection.horizontal,
+                        ),
+                      );
+                    } else if (element.type ==
+                        StringContainsElementType.hashtag) {
+                      ScaffoldMessenger.of(context).showSnackBar(
+                        SnackBar(
+                          content: Text(
+                            '${element.value} is a hashtag!!',
+                          ),
+                          duration: const Duration(seconds: 2),
+                          dismissDirection: DismissDirection.horizontal,
+                        ),
+                      );
+                    } else if (element.type ==
+                        StringContainsElementType.mention) {
+                      ScaffoldMessenger.of(context).showSnackBar(
+                        SnackBar(
+                          content: Text(
+                            '${element.value} is a mention!!',
+                          ),
+                          duration: const Duration(seconds: 2),
+                          dismissDirection: DismissDirection.horizontal,
+                        ),
+                      );
+                    } else if (element.type ==
+                        StringContainsElementType.email) {
+                      ScaffoldMessenger.of(context).showSnackBar(
+                        SnackBar(
+                          content: Text(
+                            '${element.value} is an email!!',
+                          ),
+                          duration: const Duration(seconds: 2),
+                          dismissDirection: DismissDirection.horizontal,
+                        ),
+                      );
+                    } else if (element.type == StringContainsElementType.url) {
+                      ScaffoldMessenger.of(context).showSnackBar(
+                        SnackBar(
+                          content: Text(
+                            '${element.value} is a url!!',
+                          ),
+                          duration: const Duration(seconds: 2),
+                          dismissDirection: DismissDirection.horizontal,
+                        ),
+                      );
+                    } else if (element.type ==
+                        StringContainsElementType.phoneNumber) {
+                      ScaffoldMessenger.of(context).showSnackBar(
+                        SnackBar(
+                          content: Text(
+                            '${element.value} is a phone number!!',
                           ),
                           duration: const Duration(seconds: 2),
                           dismissDirection: DismissDirection.horizontal,
@@ -179,9 +244,9 @@ class _MyHomePageState extends State<MyHomePage> {
                   highlightWordsStyle: const TextStyle(
                     color: Colors.red,
                   ),
-                  types: const [
-                    StringContainsElementType.email,
-                    StringContainsElementType.url,
+                  types: const <StringContainsElementType>[
+                    // StringContainsElementType.email,
+                    // StringContainsElementType.url,
                     StringContainsElementType.phoneNumber,
                     StringContainsElementType.hashtag,
                     StringContainsElementType.mention,